fit_garment/requirements-secondary.txt
aiohttp
basicsr
accelerate
cloudpickle
numpy==1.26.2
scipy
scikit-image
opencv-python
pillow
diffusers==0.20.2
transformers==4.33.2
xformers==0.0.20
tqdm
einops
onnxruntime
av
fvcore
omegaconf==2.1
hydra-core
pycocotools
git+https://github.com/facebookresearch/detectron2.git
git+https://github.com/facebookresearch/detectron2@main#subdirectory=projects/DensePose
virtualtryon_api_common
python-decouple
python-dotenv
huggingface_hub==0.25.2
